Restructure support/ directory
