+++ spec/acceptance/support/helpers/gitlab_signin_form.rb
