+++ spec/acceptance/support/helpers/simp_gitlab_beaker_helpers.rb
