+++ spec/acceptance/support/helpers/sut_web_session.rb
